Introduce virtual scrolling for the table on the second level
The virtual scrolling comes directly from the Angular team in the Angular CDK. For a table with a few thousands of items, the virtual scrolling will keep the items in memory and render on the screen only the ones that are visible in the viewport.

I tested the second level with a generated file consisting of 6000 entries, and everything seemed to work well.

Let me know if there are still slowdowns when working with large collections.